add a more general version of ItemsList
